+++ urlshortener/extensions.py
@@ -1,0 +1,5 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
+
+db = SQLAlchemy()
+login_manager = LoginManager()
